Fix clonks and objects falling off airships on vertical movement (#1391).
The sub-pixel position adjustment on SolidMask movement caused much smoother visuals on objects on an airship, but at the cost of things falling down because object vertices wouldn't cross the sub-pixel barrier to the next pixel value at the same time as the SolidMask.

The smooth movement is kept for horizontal.

Also fix a bug that caused objects to be added to the SolidMask movement list twice.