Fix initial object selection and cycle order when picking up objects.
The check for initial selection was broken because GetIndexOf does not return nil when the index is not found. It returns -1.

The previously intended logic (that never worked) of selecting in direction of the clonk view first may sound intuitive, but is weird in practice when you're standing just on top of an object that is 1px behind you and it selects an item very far away instead. It's particularly counter-intuitive when picking up objects out of the earth while scaling.