Lantern: Fix attachment on CreateContents, fix scenario saving and remove lantern dummy.
RejectCollect is not called on CreateContents or other forced entry methods but should still cause the lantern to attach. So move any side-effects to Collection2 instead of direct RejectCollect.

The lantern dummy had Timer=1 and wasn't doing anything. Use a simple overlay instead.

Note that interaction menus still show the lantern always with off-material. But that's a bug in the interaction menu.