Revert "Add Scale parameter in materials"
This reverts commit 8ef1f908cc638cfb8a74095c339622c8fe8cca8d. This seems to be
a performance bottleneck on some graphics cards and/or drivers. Clonkonaut
reports issues with a Radeon HD 7800.

This feature is not vital since textures can always be drawn such that they
tile properly at 512x512. While that particular problem with the Radeon card
could probably be worked around, it's just not worth it.