bucket: removed ability to place earth chunk
It was rather random whether you would place a chunk of earth or just spill the bucket. And because the spilling yielded so much more earth, placing a chunk was a waste most of the time. If we want placeable earth (which might be a good thing), we should add it to the hammer.