renamed Trunk2 to WalkableTrunk and moved it to Environment.ocd
The "Trunk2" had nothing at all to do with the other "Trunk". It's not a tree and can't be chopped. It's not a decorative object either. It's something that the scenario designer most likely has to place on purpose.
The Vegetation.ocd folder should be kept for plants and plant-like stuff that e.g. either grows or can be chopped like a tree and fulfill not another core-role (like being a decorative bridge).

The texture can possibly be scaled down a bit. But because the WalkableTrunk has a width of 264 (x*4 = 1056), I didn't want to change that now without ingame testing.