Disable Travis CI
When JCaesar dropped all Boost fallbacks, this also meant that we
wouldn't be able to build on Ubuntu 12.04 LTS anymore (because that
comes with an ancient libstdc++ which has a broken <regex>). Until we
update the Travis config to use their newer container-based build
infrastructure, there's no point having them build everything when we
already know it'll fail.