goal menu: do not have pointers to objects in the menu layout anymore (#1420)
This actually led to an infinite recursion in Mine Rescue when clicking on the goal. The goal menu would link to the goal itself which would link to the elevator -> case -> elevator -> case...
Now only layout stuff is in the layout proplists.

I noticed that with multiple goals one always gets a background. But I guess that had already been the case beforehand? Anyway, that's another issue.