fixed dangling custom overlays
Custom overlays would only be closed when the object was changed. This was not apparent for the loam, because the new overlay would always completely obstruct the old one.
But with the dynamite box displaying the sticks, the old ones never disappeared.

Now the custom overlay is explicitely closed and the update is also done if the slot previously had a custom overlay and no has none.