sawmill: hide inventory and display remaining wood to be ejected somewhere else (#1481)
The sawmill splits up objects on Collection(). That means you could put trees into the sawmill and even without power could you then take out the wood (that was yet to be ejected).
This is solved by this commit as you cannot take objects out of the sawmill manually anymore. The sawmill still displays the wood that is to be produced (because I found that cool).