Sawmill now independent from Producer library.
This was hacky to begin with and often caused problems. Lately it made the saw blade spin down after each piece of wood was cut which was something I had fixed before. This won't happen anymore.
I also made the sawing a little bit nicer by adding an engine sound. I WILL MOVE THIS TO SOUND.OCG WHEN ITS DONE