mac: fix mousewheel event handling (#1575)
Some sort of smooth scroll information was passed down to the engine which
didn't handle it properly, or the units with which they were reported to us
were different from what the engine expected.  For now, just use fixed scroll
offsets until we implement proper delta-y-aware scrolling in C4MouseControl.

The fix for #1574 might have also helped with this in recognizing when Control
is pressed and when it isn't.