fix incorrect clipping in script GUI windows (#1592)
I am not sure why the 'isMainWindow' was there. It wasn't there in 7319f7b3cc5f1ffb1fea28fd7ca84b482c6052ff and got introduced in the major rework in 049088be78db99b3aa6833847801c8e8c13ae2d7 - I guess it was just an oversight that was not noticed because usually the UI windows have the text or other things that need to scroll on a deeper level.

Anyway, checking whether the window is a script-root window does not make any sense as far as I can see now.