interaction menu: disallow interaction with outside world when contained (#1541)
The interaction with IsContainer objects is still flawed in a way: when a lorry is inside a (IsContainer-)hut, you can take the lorry into your hands. Since we don't have that possibility in the normal game, I leave it unchanged for now.

However, it is now possible to add a Chest (or a cupboard!) into a WoodenCabin and then store stuff inside of it.