ObjectInteractionMenu: Extract functions for transferring items
Transfer of items from one container to another is handled differently between the following actions: a) select items for transfer in the inventory box; b) transfer all items with the transfer-all-arrows.

This should be made consistent, especially since action a) seems to have an unnecessary special case (marked with TODO).

As a first step the methods were simply extracted.